requirements.txt
mkdocs-material
requests>=2.25.1
.
